商品追加と商品編集のバリデーションとコールバックメソッドを共通化
